Phase 8: passwd_protected codegen + cprman-driver Tier-A conversion
Lifts the BCM2711 silent-drop-on-missing-PASSWD bug class out of
driver discipline and into the framework. Drivers express field
intent; the generated `set_*` / `modify_*` accessors OR the
PASSWD byte (0x5A in bits[31:24]) into every write.

What lands
----------

- xtask/src/gen_regs.rs: lift the panic guard on `RegisterKind::
  PasswdProtected`. Refactor write emission through a `wrap_write`
  closure that produces the PASSWD-OR for passwd_protected
  registers (or `to_be()` for BE). Both the raw `write_*` and the
  typed `set_*` / `modify_*` paths route through it — drivers
  cannot accidentally bypass PASSWD by reaching for `modify_`. The
  refactor produces byte-identical output for pl011 / fwcfg /
  virtio_mmio (verified by `gen-regs --check` post-refactor); only
  cprman.rs is new in this commit.

- Validation guards on passwd_protected: u32 width only (PASSWD
  lives in bits[31:24] — narrower widths would need different
  mask/shift logic, defer until a real device needs it); rw/wo
  access only (passwd is a write-time concern, doesn't compose
  with trigger/w1c semantics); BE+passwd rejected (no current
  device combines them). Error messages name the conflict and the
  defer rationale.

- Generated tests per passwd_protected register: PASSWD wrap on
  zero payload (asserts top byte = 0x5A, low 24 bits = 0), payload
  preservation (asserts 0x00AB_CDEF input survives unchanged
  alongside PASSWD), and modify_ closure path (seed memory with
  payload, run modify that touches low bits, assert PASSWD still
  present). Three tests × 2 cprman regs = 6 wrap tests added.

cprman regspec
--------------

- `user/regspecs/cprman.toml`: flipped `emit = true`; complete
  field set on both registers (cm_emmc2_ctl: src/enable/kill/busy
  with PllDPerCore enum variant; cm_emmc2_div: divi/divf split).
- `verify_against = "lockjaw_types::clock::cprman"` +
  `[[verify_offsets]]` bindings — const_assert_eq! pins generated
  offsets to host-side constants. Moving a constant in either
  side breaks the build.
- `lockjaw_types::clock::cprman::{CM_EMMC2CTL, CM_EMMC2DIV}`
  changed from `usize` to `u64` so the cast in the generated
  const_assert (`offset_of!(..) as u64`) compiles. Comment
  documents the constraint.

cprman-driver Tier-A conversion
-------------------------------

The second Tier-A escape-valve driver (after Phase 6 ramfb), and
the first to be a clock provider rather than a display/transport.
Driver crate carries `#![deny(unsafe_code)]` with zero
`#[allow(unsafe_code)]` attribute lines; zero `unsafe` blocks;
zero `ptr::read_volatile` / `ptr::write_volatile`; zero raw
`sys_*` calls outside the `sys_exit` allowlist.

- Boot stub: `boot_stub! { hash = LOCKJAW_SOURCE_HASH, main =
  cprman_entry }`. Not `driver_main!` because cprman has no IRQ —
  the standard helper would call `bind_irq` unconditionally.
- Bootstrap + claim: `driver_bootstrap()` → `probe_by_hash()` →
  `claim_typed::<Cprman>()`. Both probe and claim Errors are
  graceful: on QEMU virt the DTB has no `brcm,bcm2711-cprman`,
  so the driver serves NotSupported for every clock op while
  the IPC plumbing stays exercised end-to-end.
- MMIO: all writes go through generated typed accessors —
  `regs.set_cm_emmc2_ctl(CmEmmc2Ctl::default().with_kill(true)
  .with_src(CmEmmc2CtlSrc::PllDPerCore))` etc. PASSWD applied
  by codegen. No hand-written `pwd()` helper, no raw `cm_write`,
  no manual KILL/ENABLE bit constants.
- BUSY poll: typed accessor `regs.cm_emmc2_ctl().busy()` in a
  bounded loop (1M iterations → ClockError::Hardware). Right
  precedent for "wait for hardware" patterns.

lockjaw-userlib::clock — server side added
------------------------------------------

Mirror of `lockjaw_userlib::clock::ClockClient` (the existing
IPC client). New `ClockEngine` trait + `run_clock_server` helper
own the receive/reply loop and error-code encoding; drivers
implement `ClockEngine::dispatch(op, clock_id_raw, arg) ->
Result<u64, ClockError>` with per-leaf MMIO sequencing only.

Third instance of the trait + server-loop shape (BlockEngine +
run_block_server, EventEngine + run_event_server, now ClockEngine
+ run_clock_server) — pattern lock-in is set.

Acceptance gates (all met)
--------------------------

1. Zero `#[allow(unsafe_code)]` attribute lines in cprman-driver
   source. Acceptance grep:
   `grep -rEn '#!?\[allow\(unsafe_code' user/cprman-driver/src/`
   returns nothing.
2. Zero `unsafe` blocks, zero `ptr::*_volatile`, zero raw `sys_*`
   outside `sys_exit`.
3. `cargo xtask gen-regs --check` clean — no incidental drift in
   fw_cfg.rs / pl011.rs / virtio_mmio.rs (wrap_write refactor
   preserves existing emission shape).
4. `cargo test -p lockjaw-regs`: 37 passed (was 25; +12 from
   cprman incl. PASSWD wrap tests).
5. `cargo test -p lockjaw-types`: 933 passed.
6. `cargo test -p lockjaw-userlib`: 28 passed.
7. `make test`: 100/100 on both GICv3 and GICv2.

Review round (Codex-paired)
---------------------------

Two follow-ups flagged for tracking (not Phase 8 blockers):

1. Tier-A boilerplate now duplicated across ramfb + cprman (boot
   → server_ep check → probe → claim). Rule-of-two threshold
   passed; a `standard_init_no_irq<T>` helper would collapse each
   to one match arm. Land before Phase 9 so the third Tier-A
   driver inherits the right shape automatically.
2. Per-field access mode missing in regspec/wirespec. The
   generated `CmEmmc2Ctl::with_busy(bool)` setter is meaningless
   (BUSY is hardware-RO) — discipline-based, exactly the kind of
   pattern Phase 8 is meant to retire one layer up. Codegen wants
   `access = "ro"` per-field, defaulting to rw. Tracked in
   docs/tech-debt.md.

Cosmetic observation (Phase 10 lockdown question): three drivers
now take a direct `lockjaw-types` dep for family constants. Should
`lockjaw-userlib` re-export the family modules so the CLAUDE.md
"drivers consume lockjaw-userlib, period" gate also covers types?
Defer to Phase 10 ergonomic-pass.

+++ lockjaw-types/src/clock/cprman.rs
@@ -64,13 +64,18 @@ impl ClockId {
 pub const CM_PASSWORD: u32 = 0x5A;
 
 /// CM_EMMC2CTL register offset (Linux: BCM2835_REG_CM_EMMC2CTL).
-/// Bits: 7 ENAB (gate), 4 KILL, 4 BUSY (RO), 4 SRC (parent select).
-pub const CM_EMMC2CTL: usize = 0x1d0;
+/// Bits: 7 BUSY (RO), 5 KILL, 4 ENAB (gate), 3:0 SRC (parent select).
+///
+/// `u64` so the regspec codegen's `verify_against` const_assert
+/// (which casts `offset_of` to u64) compiles cleanly. The constant
+/// is the source of truth that the generated module's register
+/// offset must match.
+pub const CM_EMMC2CTL: u64 = 0x1d0;
 
 /// CM_EMMC2DIV register offset. 24-bit fixed-point divider:
 /// bits[23:12] = integer part (DIVI), bits[11:0] = fractional
 /// part (DIVF, in /4096 units). Output rate = parent / divider.
-pub const CM_EMMC2DIV: usize = 0x1d4;
+pub const CM_EMMC2DIV: u64 = 0x1d4;
 
 /// CTL bit positions.
 pub const CM_CTL_BUSY: u32 = 1 << 7;
